Fixups (#8)
* KeyStore: Block key attestation for Google Play Services

In order to enforce SafetyNet security, Google Play Services is now
using hardware attestation for ctsProfile validation in all cases, even
when basic attestation is selected. The SafetyNet API response from GMS
will report that basic attestation was used, but under the hood,
hardware attestation is always used regardless of the reported state.
This results in SafetyNet failing to pass due to TrustZone reporting an
unlocked bootloader (and a partially invalidated root of trust) in the
key attestation result.

We can still take advantage of the fact that this usage of hardware
attestation is opportunistic - that is, it falls back to basic
attestation if key attestation fails to run - and prevent GMS from using
key attestation at the framework level. This causes it to gracefully
fall back to basic attestation and pass SafetyNet with an unlocked
bootloader.

Key attestation is still available for other apps, as there are valid
uses for it that do not involve SafetyNet.

The "not implemented" error code from keymaster is used to simulate the
most realistic failure condition to evade detection, i.e. an old device
that lacks support for key attestation.

Change-Id: I7282ab22b933434bb11037743d46b8a20dad063a

* SystemUI: Disable blur wakeup animation

This animation looks distracting and prevents the user from seeing
anything during the unlock animation.

Change-Id: I2c8ff9e6a9ec361595bc725d971a51e91acc13a2

* SystemUI: Add caffeine qs tile

* Caffeine Tile icon designed by Andrew Fluck

[Adin Kwok] Add long press to infinity to caffeine qs tile

Co-authored-by: Adin Kwok <adin@adin.dev>
Change-Id: If54ea27ddf179c2de39b61850163e14173727698

Co-authored-by: Danny Lin <danny@kdrag0n.dev>
Co-authored-by: Keith Mok <kmok@cyngn.com>
Co-authored-by: Adin Kwok <adin@adin.dev>